Refactor household level populations into two stages

+++ cishouseholds/derive.py
@@ -34,7 +34,7 @@ def assign_distinct_count_in_group(
     df = df.withColumn(
         column_name_to_assign,
         F.sum(F.when(F.row_number().over(count_distinct_columns_window) == 1, 1)).over(group_window),
-    )
+    ).cast("integer")
     return df
